Conflicts:
	src/main/webapp/WEB-INF/views/Header.jsp
	target/m2e-wtp/web-resources/META-INF/MANIFEST.MF
	target/m2e-wtp/web-resources/META-INF/maven/isi.project/banking/pom.properties

+++ src/main/java/isi/project/banking/model/offerInvestment/OfferInvestment.java
@@ -22,7 +22,7 @@ public class OfferInvestment {
 	@GeneratedValue(strategy=GenerationType.SEQUENCE)
 	private Integer id;
 	private String name;
-	private Integer interest;
+	private Double interest;
 	@Column(name="INTEREST_TYPE")
 	private String interestType;
 	private Integer duration;
@@ -47,10 +47,10 @@ public class OfferInvestment {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public Integer getInterest() {
+	public Double getInterest() {
 		return interest;
 	}
-	public void setInterest(Integer interest) {
+	public void setInterest(Double interest) {
 		this.interest = interest;
 	}
 	public String getInterestType() {
